Endpoints de CmpClientes readys

+++ BussinessLayer/Services/SModuloCampaña/CmpClientService.cs
@@ -1,19 +1,24 @@
 ﻿using AutoMapper;
 using BussinessLayer.DTOs.ModuloCampaña.CmpCliente;
+using BussinessLayer.FluentValidations;
 using BussinessLayer.Interfaces.IModuloCampaña;
 using BussinessLayer.Interfaces.ModuloCampaña;
 using BussinessLayer.Wrappers;
+using DataLayer.Models.Empresa;
 using DataLayer.Models.ModuloCampaña;
+using FluentValidation;
 
 namespace BussinessLayer.Services.SModuloCampaña
 {
-    public class CmpClientService(ICmpClienteRepository cmpClienteRepository, IMapper mapper) : ICmpClientService
+    public class CmpClientService(ICmpClienteRepository cmpClienteRepository, IMapper mapper, IValidateService<CmpClientCreateDto> validator) : ICmpClientService
     {
         public async Task<Response<CmpClientCreateDto>> CreateClientAsync(CmpClientCreateDto cliente)
         {
             try
             {
-                
+                List<string> errors = validator.Validate(cliente);
+                if (errors != null && errors.Count > 0) return Response<CmpClientCreateDto>.BadRequest(errors);
+
                 CmpCliente cmpCliente = mapper.Map<CmpCliente>(cliente);
                 await cmpClienteRepository.AddAsync(cmpCliente);
                 return Response<CmpClientCreateDto>.Created(cliente);
@@ -27,6 +32,10 @@ namespace BussinessLayer.Services.SModuloCampaña
         {
             try
             {
+                CmpCliente? cmpCliente = await cmpClienteRepository.GetByIdAsync(idCliente, idEmpresa);
+
+                if (cmpCliente == null) return Response<string>.BadRequest(new List<string> { $"Usuario con ID: {idCliente} no fue encontrado" });
+
                 await cmpClienteRepository.DeleteAsync(idCliente);
                 return Response<string>.NoContent("Recurso eliminado");
             }
@@ -62,7 +71,7 @@ namespace BussinessLayer.Services.SModuloCampaña
 
                 if (clients != null)
                 {
-                    
+
                     List<CmpClienteDto> cmpClienteDtos = mapper.Map<List<CmpClienteDto>>(clients.ToList());
                     return Response<List<CmpClienteDto>>.Success(cmpClienteDtos);
                 }
@@ -74,9 +83,20 @@ namespace BussinessLayer.Services.SModuloCampaña
             }
         }
 
-        public Task<Response<string>> UpdateClientAsync(int idCliente, CmpClienteUpdateDto cliente)
+        public async Task<Response<string>> UpdateClientAsync(int idCliente, CmpClienteUpdateDto cliente)
         {
-            throw new NotImplementedException();
+            try
+            {
+                CmpCliente? existingClient = await cmpClienteRepository.GetByIdAsync(idCliente, cliente.EmpresaId);
+                if (existingClient == null) return Response<string>.BadRequest(new List<string> { $"Usuario con ID: {idCliente} no fue encontrado" });
+                CmpCliente cmpCliente = mapper.Map<CmpCliente>(cliente);
+                await cmpClienteRepository.UpdateAsync(cmpCliente);
+                return Response<string>.NoContent("Usuario editado");
+            }
+            catch (Exception ex)
+            {
+                return Response<string>.ServerError(ex.Message);
+            }
         }
     }
 }
